requirements.txt
boto3==1.24.10
Pillow==10.1.0
Pillow==10.1.0
